big shifting around - n.b. something will need to be done about *.xml
