--- downloadTrackers.sh
@@ -1,4 +1,0 @@
-# based on the link on the gForge instructions page
-# typically doesn't work because of google docs redirections - try
-# to find a stable way of doing this
-curl -o trackers.xml -L https://docs.google.com/a/lmckenzie.com/uc?id=0B285oCHDUr09Q2VENW9Ta005QkE
