--- get-phd-trackers.sh
@@ -1,1 +1,0 @@
-saxon -s:tracker.xml -xsl:trackerByBallot.xsl -o:phdtrackers.xml which-ballot='2018-Jan PHD'
