--- list_onedrive.sh
@@ -1,1 +1,0 @@
-ls /users/jrhoads/"onedrive - philips"/Projects/FHIRBallots
