--- put_onedrive.sh
@@ -1,1 +1,0 @@
-rsync --dry-run -r /Users/jrhoads/Projects/FHIRBallots /users/jrhoads/"onedrive - philips"/Projects/FHIRBallots
